🔥Fix: Header , Split logic
