+++ app/helpers/constants.js
@@ -1,0 +1,5 @@
+export const CategoriesEnum = {
+  movies: 'movies',
+  games: 'games',
+  music: 'music',
+};
